# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for GCW API.",
        "title": "GCW API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1/gcw/resources",
    "paths": {
...
        "/team/registration/cp": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team Registration"
                ],
                "summary": "Register CP Team",
                "parameters": [
                    {
                        "description": "Register CP Team",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegistrationCPTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RegistrationCPTeamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/registration/ctf": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team Registration"
                ],
                "summary": "Register CTF Team",
                "parameters": [
                    {
                        "description": "Register CTF Team",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegistrationCTFTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RegistrationCTFTeamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/registration/find/{join_code}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team Registration"
                ],
                "summary": "Find Team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Join Code",
                        "name": "join_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RegistraionTeamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/registration/hackathon": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team Registration"
                ],
                "summary": "Register Hackathon Team",
                "parameters": [
                    {
                        "description": "Register Hackathon Team",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegistrationHackathonTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RegistrationHackathonTeamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/registration/join/{join_code}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team Registration"
                ],
                "summary": "Join Team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Join Code",
                        "name": "join_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RegistraionTeamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.Member": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "dto.RegistraionTeamResponse": {
            "type": "object",
            "required": [
                "supervisor",
                "supervisor_nidn",
                "team_name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id_team": {
                    "type": "integer"
                },
                "leader": {
                    "$ref": "#/definitions/dto.Member"
                },
                "member": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Member"
                    }
                },
                "order_id": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "qr_string": {
                    "type": "string"
                },
                "supervisor": {
                    "type": "string"
                },
                "supervisor_nidn": {
                    "type": "string"
                },
                "team_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.RegistrationCPResponse": {
            "type": "object",
            "properties": {
                "bukti_pembayaran": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "domjudge_password": {
                    "type": "string"
                },
                "domjudge_username": {
                    "type": "string"
                },
                "id_cp_team": {
                    "type": "integer"
                },
                "id_team": {
                    "type": "integer"
                },
                "join_code": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "qr_string": {
                    "type": "string"
                },
                "stage": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dto.RegistrationCPTeamRequest": {
            "type": "object",
            "required": [
                "supervisor",
                "supervisor_nidn",
                "team_name"
            ],
            "properties": {
                "bukti_pembayaran": {
                    "type": "string"
                },
                "join_code": {
                    "type": "string"
                },
                "supervisor": {
                    "type": "string"
                },
                "supervisor_nidn": {
                    "type": "string"
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "dto.RegistrationCPTeamResponse": {
            "type": "object",
            "properties": {
                "cpteam": {
                    "$ref": "#/definitions/dto.RegistrationCPResponse"
                },
                "team": {
                    "$ref": "#/definitions/dto.RegistraionTeamResponse"
                }
            }
        },
        "dto.RegistrationCTFResponse": {
            "type": "object",
            "properties": {
                "bukti_pembayaran": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id_ctf_team": {
                    "type": "integer"
                },
                "id_team": {
                    "type": "integer"
                },
                "join_code": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "qr_string": {
                    "type": "string"
                },
                "stage": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.RegistrationCTFTeamRequest": {
            "type": "object",
            "required": [
                "supervisor",
                "supervisor_nidn",
                "team_name"
            ],
            "properties": {
                "bukti_pembayaran": {
                    "type": "string"
                },
                "supervisor": {
                    "type": "string"
                },
                "supervisor_nidn": {
                    "type": "string"
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "dto.RegistrationCTFTeamResponse": {
            "type": "object",
            "properties": {
                "ctfteam": {
                    "$ref": "#/definitions/dto.RegistrationCTFResponse"
                },
                "team": {
                    "$ref": "#/definitions/dto.RegistraionTeamResponse"
                }
            }
        },
        "dto.RegistrationHackathonResponse": {
            "type": "object",
            "properties": {
                "bukti_pembayaran": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id_HackathonTeam": {
                    "type": "integer"
                },
                "id_team": {
                    "type": "integer"
                },
                "join_code": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "qr_string": {
                    "type": "string"
                },
                "stage": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.RegistrationHackathonTeamRequest": {
            "type": "object",
            "required": [
                "supervisor",
                "supervisor_nidn",
                "team_name"
            ],
            "properties": {
                "supervisor": {
                    "type": "string"
                },
                "supervisor_nidn": {
                    "type": "string"
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "dto.RegistrationHackathonTeamResponse": {
            "type": "object",
            "properties": {
                "hackathonTeam": {
                    "$ref": "#/definitions/dto.RegistrationHackathonResponse"
                },
                "team": {
                    "$ref": "#/definitions/dto.RegistraionTeamResponse"
                }
            }
        },
...
        "helper.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}